move button color into helpers dir.
