+++ lib/src/helpers/auth_button_color.dart
